requirements.txt
typer>=0.9.0
rich>=13.0.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4 